net/netcheck: adjust HTTPS latency check to connection time and avoid data race
The go-httpstat package has a data race when used with connections that
are performing happy-eyeballs connection setups as we are in the DERP
client. There is a long-stale PR upstream to address this, however
revisiting the purpose of this code suggests we don't really need
httpstat here.

The code populates a latency table that may be used to compare to STUN
latency, which is a lightweight RTT check. Switching out the reported
timing here to simply the TLS connection setup time avoids the
problematic package, and provides timing results that are more
comparable. Retaining the HTTP request to ensure that the node is
actually reachable still has some value in demonstrating that we could
use the node.

Fixes tailscale/corp#25095

Signed-off-by: James Tucker <james@tailscale.com>